Setup service sign in format boilerplate
Adds only the choose sign in files for now.

Rails looks in "app/views" for a template that matchs the format
type. As that is to a one-to-one relationship for the service_sign_in
format, we have created a base template ("service_sign_in"), that will
then call a partial template, based on the page that the user is on.