Split tasklist title styles to remove margins
The margin on tasklist titles isn't suitable for universal layout,
but we need to ensure that Answers, Guides and Help Pages preserve
the original margin as they will be moved to the new layout after
publications.
I've added a fixme to remove the temporary styles once all formats
are moved to the new layout.