Moves Tasklist functionality to GovukNavigationHelpers
This commit is using the github repo alphagov/govuk_navigation_helpers with
`add-tasklist-functionality` branch. This is meant to be deleted after
the PR is merged and people are happy during review.

We are moving the Tasklist functionality into GovukNavigationHelpers so
we can consolidate the code in one place.

This PR depends on the release of a new version of
GovukNavigationHelpers with the following PR
https://github.com/alphagov/govuk_navigation_helpers/pull/90

Trello: https://trello.com/c/JiPFogRf