Show "or" in radio button list in choose_sign_in template
We need to group the radio options based on the type of option
being presented to the user.

For example

* Sign in option 1
* Sign in option 2

or

* Create new account

At the moment we don't have set rules about which items will be grouped
together, so for now, the last option in list goes under the "or".

This logic will be updated when the rules are agreed with content designers.