Ensure publication_metadata doesn't contain national applicability
National applicability (aka 'Applies to' links) are present
in metadata and publisher metadata while we switch designs for
specific formats. So remove it from publisher_metadata for
publications while it is being moved to the new layout.
Once all formats have moved it can be removed from publisher
metadata completely.