Moves Tasklist functionality to GovukNavigationHelpers
We are moving the Tasklist functionality into GovukNavigationHelpers so we can consolidate the code in one place.

It no longer uses `concerns` and instead the functionality is split in
two objects:

- TasklistContent: receives the current request.path and tries to find a
  tasklist that has that path, in case of one being used between "End
Civil Partnership" and "Get Divorce", "Get Divorce" tasklist is the
default.
- CurrentTasklistAbTest: Depends on having a current tasklist set to
  enable the AbTest and set it on the response headers.

This PR depends on the release of a new version of GovukNavigationHelpers with the following PR
https://github.com/alphagov/govuk_navigation_helpers/pull/90

Trello: https://trello.com/c/JiPFogRf

![screen shot 2017-12-08 at 10 55
10](https://user-images.githubusercontent.com/136777/33766082-cd37fafe-dc13-11e7-8643-3e5bb53c3b86.png)
![screen shot 2017-12-08 at 10 55
24](https://user-images.githubusercontent.com/136777/33766083-cd52d306-dc13-11e7-89a5-3235a27f8b1b.png)