Turn off CSRF protection on service_sign_in POST route
When we tried submitting a radio button on the choose-sign-in page the
following error is thrown:

`Can't verify CSRF token authenticity.`

As we are not submitting and sensitive data, and are just redirecting users,
we don't need this protection on the form.

We don't currently have a session store due to the layers of caching, so we
did not try to set the authenticity token, and have instead disabled it.