Make publisher metadata styling consistent
When moving publications to the new layout we wanted
to combine the publisher metadata with an optional
logo and make the combination span all 3 columns.
This change needs to be reflected in formats which
won't have logos but will have a full width publisher
metadata border.
I've made a partial rendering the metadata component
in a 2/3 width column and the logo in the 1/3 column
because it seemed the most reasonable way to combine
both without compromising the component design.