Render translations correctly on publications
Because of the way we use template variants, when we call the
shared/translations partial it renders the
shared/translations+taxonomy_navigation variant. Publications that are
tagged to a new taxonomy use this variant to show an alternative
breadcrumb.

However, the taxonomy_nav version of the partial expects to be inlined
in content, so doesn’t have a grid class wrapping it.

Because of Rails magic loading in this variant the code looks fine, and
it only breaks on certain publications which meat the taxonomy nav
criteria.

* Don’t use the shared translations template

This duplicates some code which can be DRY’d up when we no longer need
the taxonomy_nav variants.