Remove wildlife package test
We need to remove this check as we are renaming the field in content schema
and if we leave this value in we are unabel to get a green build.

We will investigate the value of adding this check back in with the update
link text once we have completed the content schema change.