Show contents list depending on content character count
We have found instances where displaying a contents list on a page isn't always
appropriate, such as when there is only one item in the list. We have decided
to not show the contents list if there are fewer than three items, with an
additional condition that we do show the contents list if the first item's
content is more than 100 characters.