Remove redundant conditionals for publishing components
We are always requiring the `govuk_publishing_components` gem, so
`GovukPublishingComponents` will always be defined, making these
conditionals always true.

We used to specifiy the gem with `require: false`, but that's no longer
the case since we're using the gem to ship the task list component
(https://github.com/alphagov/government-frontend/pull/578).