Add link to finder related to this content page
Many formats rendered in government-frontend can
be filtered via a faceted search on a finder page.
Linking to this finder may provide the user a simple
route to finding content similar to, or related to
the content page they are viewing.
So provide a contextual link back to the appropriate
finder page, with facets populated with the relevant
organisation, person, world_location and policy.
We will track the usage of these links and decide on
their efficacy once we have a decent amount of data.

The formats affected are announcements, publications,
consultations, specialist documents and statistics.