Update Working Groups integration tests
We update test coverage to reflect that contents lists are shown under certain
conditions.

We have found that Working Groups with policies don't always have a body, which
caused problems in the current code implementation when it came to checking
content length in the first item of a contents list. Therefore we now check if
a "first item" is present in the body before carrying out the content checks.