Updates for navigation_helpers/pub_components changes
govuk_navigation_helpers no longer contains AB testing logic for enabling/
disabling step by step navigation.  It also has a change to the format of the
data to pass to the task_list component.

govuk_publishing_components has the corresponding change to the format of the
data that task_list components expect, but also has renamed task_list elements
to step_by_step_nav.

We need to update this code to work with that.  It makes for an unwieldy review
but nothing will work if we do piecemeal commits. :-/