Add first_item to document_collection presenter
We have found instances of document collections that don't have a `body` in
`details`, which meant that the previous implementation of checking for the
first item in a contents list didn't work for this type of content. We add a
new `first_item` method for document collections which, in the absence of
`body`, gets the body of the first collection group. It then returns the first element of the parsed body.

We also update the `first_item` method in the `ContentsList` module to return
the next element after a h2 element, as the previous implementation wouldn't
work for document collections where we don't need to skip the h2.