Add guards to Contents List
Now that it's possible to check through a Document Collection body or
collection group body which does not contain a h2 tag, we found that this
sometimes causes errors in our first item checks which loop through until a h2
is found. This adds some guards to break the loops when there are no more
elements to check.