Add first_item method to Working Group presenter
Working Groups don't always have h2 tags in their body, which has caused a
problem in the `first_item` method in the `ContentsList` module which looks for
a h2 element in a content item's body. We add a `first_item` method to the
Working Groups presenter to check for the presence of an h2 tag in the body,
and if it is absent return the first element in the body.

While solving this problem we have found an edge case for Working Groups in one
of the integration tests which uses a content item example that doesn't have a
body. We don't know if this is common across Working Groups but we have added a
guard to check for the presence of a body before looking for HTML elements.