Render European structural investment funds logo
Publisher metadata should render the relevant logo next to it,
previously this worked for National Statistics but the European
structural investment funds logo was not tested and therefore
not included in the move to a new layout.
This commit adds a presenter unit test and a format specific
integration test to cover the rendering of this logo.