Add support for active page
Guide chapter headings are changing to match the format of Content
Lists and this change supports that. The main difference is that the
guides have contents that span different urls and it isn't desirable
to link a page directly to itself. This adds support for a list item without a link
in a contents list.