Add support for blank hrefs
Guide chapter headings are changing to match the format of Content
Lists. Rather than creating a new component it makes sense to
reuse this one. The main difference is that the guides have contents
that span different urls and it isn't desirable to link a page
directly to itself. This adds support for a list item without a link
in a contents list.