AB test guide chapter navigation
We want to test chapter navigation which does not use numbers, as we think this
may not aid comprehension and may be confused with step by step navigation.

We're using the [contents-list component](https://government-frontend.herokuapp.com/component-guide/contents-list)
as the alternate, because it fits our needs and is already used in other formats.

We are not including Education format as there are other tests inbound for this
and we don't want to affect them.