Disable caching of content-store responses
By default, gds-api-adapters caches responses based on the `max-age` set
in the response's cache control headers. This means that
government-frontend can use a very stale max-age when setting cache
headers on its own responses.

In the worst case scenario, this means that content changes can take up
to 30 minutes to appear on the site.

Disabling this cache should not put much more load on the content store
because responses are cached so heavily in Varnish and the CDN, and
these handle cache headers correctly.