Update presenter usage of related navigation
The local version of `related-navigation` component relied on related
navigation data being passed to it from a content item presenter. The
`related-navigation` component in `govuk_publishing_components`, which replaces
the local version, accepts a whole content item and does the job of figuring
out what data should be in related navigation, so we no longer need the
`related_navigation` method in the presenters to provide this.