Only show taxonomy sidebar when tagged to a live taxon
The recent changes to the taxonomy exposed that we have duplicated
logic between this application and the components. At the moment this
app thinks that there are taxons, so it tries to show the sidebar.
However, the `govuk_component/taxonomy_sidebar` component *doesn't*
think there are taxons, because it only looks at the live ones.

This is a temporary fix, we will re-architect this so that the logic is
simple and in only one place.