Stop using request variants to determine layout
This removes the code for the taxonomy_navigation requests variants and
uses `if` statements in the views instead.

This makes it clearer what's going on. It shouldn't make any change in
the pages.

I suspect that most of these `if` statements aren't actually meaningful
and can be removed. That's for a follow up PR.