Use contextual navigation components
This PR uses the contextual navigation sidebar and breadcrumbs from the
publishing components gem. It removes all the logic for what to display
where and replaces it with calls to the components.

The intention is that this PR doesn't change the layout, but there are
a few changes:

- Contacts have the new sidebar, but the data has stayed the same. I'm
unsure why universal layout hadn't been implemented for contacts in the
first place, and Ganesh wasn't sure either. This might have been an
oversight.
- On topical event pages about pages, there's no more "(archived)" in
the title when the topical event is archived. I think this is an
acceptable trade-off given the complexity it would mean moving this
special functionality to the components.
- We no longer override the header navigation type for things tagged
tagged to the taxonomy. I don't think there was very much reason behind
showing it and not showing it, so I'm okay with this change (this is
the content_for :simple_header stuff that's removed)
- Even though there's a related_items method in the travel advice
presenter, it was unused and the travel advice pages are unchanged
after this PR.

https://trello.com/c/1LEofaHC