Replace publication header with inverser header
We created a new inverse header component as part of the work for building new topic pages. This component is a shared component in the govuk_publishing_components gem. As it behaves very similarly to the existing publication header component in government-frontend, it makes sense to try and consolidate the two.

This commit therefore replaces uses of the publication header with inverse header, and removes the publication header component.