Set title lang to match locale of content
See https://github.com/alphagov/govuk_template/issues/342

Currently we set the html level lang to "en" and add a lang
attribute to the "main" element with the locale of the content.

This causes screen readers to interpret the page title in
english. We should instead markup the title with the same
langugage (both the title and main content are in the
content item's language).