Add structured data for news articles
This implements a minimum of structured data for news articles.

 As per Google:

> Structured data is a standardized format for providing information
about a page and classifying the page content; for example, on a recipe
page, what are the ingredients, the cooking time and temperature, the
calories, and so on.

https://developers.google.com/search/docs/guides/intro-structured-data

We use the JSON-LD (JSON "Linked Data", https://json-ld.org) here. It's
Google's recommended way of implementing this and avoids a lot of the
hassle involved with Microdata, which interleaves the structured data
with HTML.

The output of this is tested with the Structured data testing tool:

https://search.google.com/structured-data/testing-tool

Unfortunately, the "publisher" attribute needs a logo. I've chosen the
Opengraph image for this because it's not likely to change and will
probably stay around for a long time. If we're going to implement more
structured data, we'll have to come up with an alternative (and a
better image).