HUB-143: Small style change to description text on service sign in pages
An option to display a text paragraph description under the heading on the service sign in pages has existed for some time but was not being used. We are about to start a base line test to see if having a description here improves through put by increasing understanding. However when testing what this would look like we found that the text was a little small and close to the error text (when shown). This commit contains small changes to change the text size and element spacing.

solo: @rachelthecodesmith