Add missing structured data
This adds the baseline structured data to all remaining formats in the
app.

Everything is an article, except speech and
world_location_news_article, which are news articles.

I think there are a bunch of improvements we could make with this, like
marking up statistical datasets as
https://developers.google.com/search/docs/data-types/dataset, but
that's for future PRs.

https://trello.com/c/RLQI1IUV/23-add-structured-data-to-the-top-500-page
s