Make HTML publication page more robust
Since moving the organisation logo to the gem we're seeing intermittent
test failures. This happens in cases where the random content generator
in `HtmlPublicationTest` generates an organisation item without a logo
in the details. This makes [`organisation_logo` return nil][1], which
is passed into the component, which subsequently crashes with a
nil-error.

This makes the page more robust by only displaying logos that exist.

[1]:
https://github.com/alphagov/government-frontend/blob/fc0f75cd16f4e1f1266
90cf5f9aee005af504186/app/presenters/content_item/organisation_branding.
rb#L4