HUB-97: Log choice of verify users
We'd like to see what choices exisiting Verify users make on the
service sign-in page. Therefore we're calling a Verify endpoint using
an ajax call and based on the user cookie to decide whether it's a
Verify user and report accordingly. This is a baseline test to learn about the users' behaviour.

Co-Authored-by: Jakub Miarka <jakub.miarka@digital.cabinet-office.gov.uk>