Add supergroup and rummager logic
Puts in place the supergroup (services) and rummager search logic in place ready to retrieve the most popular services.

We are also stubbing the rummager requests in the tests.

The test helpers in gds-api-adapters are all set to stub rummager responses.
This was done in https://github.com/alphagov/gds-api-adapters/commit/ddae94d1656f59f7778ca794a0cce356fc75e7e2
as an explicit choice.