Fix test failures caused by browser caching
Using Chrome instead of PhantomJS means cache headers are being
correctly observed.

A number of test cases share the same content item fixture, but make
changes to the body of the content item for the specific test case. We
therefore need to add a cachebust string to the URL as the content
item base path remains the same.