Add supergroup sections to taxonomy navigation
Adds guidance and regulation, and policy and engagement sections to the nav
section at the footer of content pages.

This is currently concealed from public view by an AB test.  The navigation
is visible if you're in the B sample of the ContentPagesNav test.

The content page must itself be tagged to a live taxonomy, be rendered by this
application and not be an HTML publication.

Policy and engagement content is configured to get latest content.
Guidance and regulation is retrieved by popularity (as determined by rummager).

This PR is to get things generally close to the desired outcome and polish
after a design review.