Add tests and tracking for guidance / policy sections
Wrapping the taxonomy block in a track-click data module activates link
tracking.  I've checked this in GA debug and it seems to do the trick.

Also switching the guard clauses on each section from any? to present? as this
checks whether hashes/arrays are nil or empty.  (The services data can be nil)