Add transparency section
Add the transparency supergroup to the taxonomy navigation section.

As per the other PRs, this is currently behind an AB test and is not seen
by general access.

Transparency docs are displayed as lists of items rather than highlight boxes.

The content page will show this navigation if they are tagged to a live
taxonomy, be rendered by this application and not be an HTML publication (and
 have results to show)

You will need to have an extension like [ModHeader](https://chrome.google.com/webstore/detail/modheader/idgpnmonknjnojddfkpgkljpfnnfcklj?hl=en) installed in your browser to see the B Variant. The values should be:
Request Header Name: GOVUK-ABTest-ContentPagesNav
Request Header Value: B

https://trello.com/c/A1x3YnId/51-get-most-recent-transparency-data