Add news and communications section
News and comms are different from the other sections in that we show the image
from the article if it's a featured link (and the image is present).

This will be updated in future to get content images from rummager which
removes the need to call content store for each item.

We'll also update this to display results as one list rather than separate
items.