Remove CSS override for component
We were overriding the CSS for the heading component. It's bad practice to do this, we should instead make use of the parameters which can be passed into the component.

This will make the headings render as 19px on desktop and 16px on mobile (as opposed to 16px on all devices).