Add banner to the taxonomy navigation
This adds a breadcrumb and a banner navigation element listing things a
content item is part of.

If the content item is tagged to a single topic a full breadcrumb to it
is shown. If it is tagged to multiple topics then a home breadcrumb is
shown and the topics are listed in the banner.

Some examples:

- One topic and step by step: /vehicle-insurance
- Multiple topics and a step by step: /vehicle-registration/new-and-used-vehicles
- One topic, no step by step: /pay-dartford-crossing-charge

https://trello.com/c/rznwkADc