Catch rummager errors for navigation
We reduce the timeout for rummager queries to 2 seconds.  If they go above 2 secs
for all of the queries we risk timing out the whole page which is not good.
(They are generally well below this 99th &ile is at 300ms, 50th %ile is at 21ms
when running on staging at double our expected load)

In addition, we catch any errors returned from the api call and set the results
to empty so that we can render what we need to.

We send metrics of these errors to Graphite so we can analyse them later.