Fix spacing for notices in consultations
When we replaced this with the notice component in the gem, we removed the bottom margin in all cases.
This meant that sometimes, the heading of a section could appear directly underneath a notice with no spacing.

This fix introduces spacing in those cases, but adds a bottom margin when the notice is immediately followed by a banner item (as per the original design)