Fix content out links for content pages AB
Previous calls to rummager were always being filtered by subgroups, regardless of the supergroup. This meant some supergroups were not displaying at all.

This commit changes the retrieval of subgroups to link them to the current supergroup for that search. Only subgroups relevant to that supergroup are returned. If there are no subgroup filters for that supergroup, all content is retrieved.