Pass whole request object to publishing components
In version 9.18.0, the constructor for Contextual Navigation was changed and now expects the whole
request object to be passed. This is because Contextual Navigation also needs to check the values in
The query string (if it exists) to see if the user is on a step by step journey. If they are, the content id of
the step by step is being appended to the querystring.