Replace h4 with p styling for contacts
We have some `h4` tags on the page that are unnecessary. Replace them with
`p` tags instead.

Contact pages on goverment-frontend can be represented by the following
pseudo code:

<h2>Phone</h2>
  <h3>Phone group title</h3>
    <p>Telephone number</p>
    <h4>Opening times</h4>
    <h4>Best time to call</h4>

  <h3>Phone group title</h3>
    <p>Telephone number</p>
    <h4>Opening times</h4>
    <h4>Best time to call</h4>

Some contact pages, however, do not have multiple "Phone groups". In
these cases, the "Phone group title" `h3` is not displayed (there is
specific logic to only show the titles if more than one exists). When we
do not display the `h3` on the page, we end up with a heading hierarchy
that skips from an `h2` (Phone) to an `h4` (Opening times).

To add insult to injury, the `h4` renderd markup has
`class="normal-weight"` to override the default bold behaviour of `h4`
tags.

There is very little need for such an explosion of heading levels and
the content still makes sense even without the `h4` tags.