Explicitly precompile assets on test
When Capybara first runs it triggers an asset compilation. This takes a
very long time at the moment and causes Net::ReadTimeout errors when
merged to master. This commit adds an explicit precompile to the test
helper so that assets are already compiled by the time Capybara does its
thing. This will slow down testing on dev environments when first
running tests.