Include precompile assets  argument to Jenkinsfile
When Capybara starts during our tests, it causes assets to precompile.
This can be quite slow and often the first request times out during our builds when
merging to master while waiting for asset compilation.

We want to precompile assets before running the tests so we don't hit
this timeout issue and fail our buid.

Pass the precompileAssetsBeforeTests argument to govuk.buildProject
which will trigger a precompile before running tests.