Add additional variant for Related Links A/B test v3
This commit adds an additional `C` variant for the upcoming related links A/B test iteration #3. In the previous test we saw some users who had the control variant cookie, yet navigated via the related links which are only shown on the `B` variant.

In order to check if there are any problems with our A/B testing framework, in addition to making the data collection easier for later analysis, we are introducing the new `C` variant which will become our test variant. How the variants now work are as follows:

- `A` variant: No users should ever end up in this variant, as our CDN will be configured it to assign 0% of users. If users do end up in this variant, we may have issues with our A/B testing framework which will require further investigation.
- `B` variant: Our control variant, where related links will not be changed in any way.
- `C` variant: Our test variant, where related links will be changed to use our generated links for the sample involved.