Add new locales
We are adding support for:

Croatian: hr
Danish: da
Dutch: nl
Finish fi
Icelandic is
Irish gd
Norwegian: no
Maltese: mt
Slovenian: sl
Swedish: sv

Add the locale codes to `config/application.rb` to enable them and also add
the files to `config/locales`. No translations are currently provided in
the locale files. These will be added at a later date. For now, we just
need to enable adding translations to publications for these languages.

Note that in config/locales/en.yml we've enclosed "no" in quotes. This
is because in YAML no is a [reserved word](https://yaml.org/type/bool.html).
We use rails_translation_manager to regenerate our locale files.
When rails_translation_manager loads the YAML and interprets
"no" as a bool instead of a locale, it breaks.