Use request header for related links to determine response header
This commit fixes the related links response header to be determined by the request header. If the request header is not set or does not match the required value as defined by the `feature_flags` initializer, we set the response header to `false`; if the request header is set and matches the required value, it's set to `true`.