Force English for metadata
On translated pages we are largely failing to translate the metadata,
leading to messy strings where the only translated content is the month
or the word "and" amongst otherwise English sentences.

This change sets the locale to English for the dates, and reverts attempts
to translate other sections of the page, and it marks the components
as being in English. It removes classnames that aren't used from the CSS
and forces the text direction to be left-to-right for these components.

This marks the page up more semantically, and supports a better experience
for screenreader users as they can now context switch the language
appropriately.