Add detection of contextual title translation
This adds a helper method which detects whether the string requested
was returned by the current locale's translation file. If not, it returns
the default locale.

This helper is in turn used on the contextual headings, allowing us to
pass the information on to the component that displays the heading, meaning
we will be able to markup the contextual header's actual language, supporting
screenreaders and making the page more semantic.