Don't show failed translation message for missing document types
With the update to Whitehall that allows adding any GOV.UK page to a
document collection there is the potential users could add documents
with types the Government Frontend knows nothing about. For these cases
this will return a nil value rather than raise a failed translation
message.

At this point in time it seems unlikely users would use a document type
that is not known about as the expected usage is Content Publisher
document types. However there is nothing to prevent wider usage of
document types.