Ban indexing in preview apps
Adds a noindex,nofollow to pages rendered using the application layout in
Heroku review apps.

Also disallows spiders from crawling this app using the robots.txt.  The
robots.txt will only be accessible from apps that are not behind router.

This means that Heroku preview apps will indicate to crawlers that they should
not index the site, but applications in production environments will be unaffected.