Adjust focus states
This adds the govuk-link class to multiple links throughout the codebase
in order to inherit the correct styles, and updates related tests. Where
the class is overridden by govuk-template the styles have been updated to
use a temporary mixin to enforce the correct colour state and the class is
locally overridden where white links are needed.