Deduplicate FAQ page schema presentation
At present we're showing the same FAQ schema on each page of a guide. This
is a bit suboptimal.  We have the option of splitting it up and showing one
question and answer per part/chapter in the guide, but it seems much better
to present it to Google as a joined up publication, because that's how it's
been designed.

To resolve this problem, we fall back to the Article schema for chapter pages,
so we're still presenting some schema.org content, but not presenting duplicate
content.