Render a guide as a single page
We think it might really help our SEO to have guide content on one
page.  The how to vote page is particularly interesting because we have some
data about searching intents that might bring people to this page.  We also
think that amending this format to a single page might benefit voice answers.

There are some existing issues with guides having multiple h1 tags on a page.
This PR doesn't attempt to fix it (and in a way makes it worse) because we
have more h1s on the page than we used to.  I'm led to believe that this is
not relevant to WCAG by this [trello card on the GOV.UK accessibility backlog](https://trello.com/c/LzxvPjUz/117-fix-multiple-h1-heading-structure).

Other than that, this has been mostly stolen from the existing guide layout,
only including all of the parts.