Redirect chapters to anchors in page
We're testing this guide as a single page.  We want existing links to
behave nicely, so we'll redirect subchapters to the default page in the
guide.

We're also using the slug to try to find the section of the page they're
looking for.  This works because the IDs of in page headings are the same
as the slug of the chapter that they refer to.