Temp revert to previous FAQPages on Guides
This is a temporary reversion to the original implementation of FAQPage
schemas on Guides.  We're running user research next week, and want to
use this version's Google results.  If the research goes well, we'll make
a further decision then whether to roll this into the standard implementation.

It was [originally implemented in govuk_publishing_components](https://github.com/alphagov/govuk_publishing_components/pull/1087)
but I don't want to revert it in publishing components yet because:

- Answers also use that implementation, and I don't want to mess with them right now
- This version is temporary until 31/10/2019 when we'll decide what to do

This version of the schema includes the entirety of a guide's content in the FAQPage
schema on the first chapter.  The subsequent chapters do not include an FAQPage
schema to avoid duplicating rich results (but do include the Article schema).