Use heading component instead of h1 tag
This allows us to use standard padding, borders, styles etc.

This doesn't fix the issue of multiple h1s on the page, but this
problem can only be addressed by first shuffling down the heading
levels of the published content. At present, content within the body
starts at h2, which doesn't leave us any room for separate heading levels
for the page title and individual chapter titles.

As mentioned in the previous commit, this is not a WCAG issue though,
so I don't intend to try to fix that here.