Add hardcoded FAQPage schema for voting-in-the-uk
We think that the voting intents around this page mean that users will benefit
from a more curated schema.  This is intended to be a one of a kind for now,
and is obviously time sensitive, hence the somewhat blunt approach.

I've added a couple of light touch tests to ensure that they aren't broken
inadvertently.

We'll be keeping the content in the yaml file in step with the published
content until we find a better approach.