Update /voting-in-the-uk schema
The content was updated on Friday.  There is [a diff of the content changes](https://publisher.publishing.service.gov.uk/editions/5dbc325c40f0b670153e4e18/diff)
in Publisher.

Only certain sections of the page are included in the schema, because there's
a limit to the amount of content that can be included in the FAQ schema.
(That's also the reason for using a separate yml config file for this rather
than generating it dynamically from the content item).

The [initial version of the page](https://docs.google.com/document/d/14g7FK-v_Y5iqWcB9y5WqTUWU6JuY_eaSGAqxXSMb0Ek/edit#heading=h.9w2yx5uysegs)
(with the relevant FAQ sections highlighted) is available to review if
necessary. This provides a decent indication of which bits are included,
but is now out of date.

Speak to me or Sam Dub for further details if required.