Update voting eligibility to match live content
The diff of this is [markdown in publisher](https://publisher.publishing.service.gov.uk/editions/5dd7c2dfe5274a0aca896563/diff)
which might be easier to check the changes.

The content is now live though, and so the html should match that in the
[Eligibility to vote section](https://www.gov.uk/voting-in-the-uk?sdfsd#eligibility-to-vote)

I've added the `src=schema` query string to the links to help us monitor
traffic driven from the schema.org json.