Switch to using the government from links, rather than the details
Currently, Whitehall includes the details of the associated government
for some content in the details. Now Whitehall also links the content
to a content item representing the government.

Using the linking approach is preferable going forward, as that means
the Publishing API can take care of updating the content store when a
government changes (becoming historical), rather than Whitehall having
to publish new editions.

This change to Government Frontend can be deployed once all the
relevant content has been linked to the correct government, something
which is being done in bulk from Whitehall.