Use a placeholder image from asset-manager
In the past this linked to an image served by Whitehall, but because of
the way Rails generates hashes for assets, we can't guarantee that this
URL will always exist. This hasn't been a problem in the past because
we've been able to put an unhashed version of the image on the Whitehall
machines, but since moving to AWS the machines can be recreated at any
moment and won't have the file.

To solve this problem I've uploaded the file to Asset Manager and we can
link to that instead. Ideally, we wouldn't depend on a file in a
different application, but this solves the problem for now.