Handle old-style /goverment/uploads routes
This is currently handled by Whitehall but we're trying to move away
from Whitehall being a frontend application. The next best place for
this bit of code seems to be Government Frontend.

https://github.com/alphagov/whitehall/blob/master/test/functional/asset_manager_redirect_controller_test.rb

The code itself is used to handle assets which are still being requested
on their own URL before we migrated all assets over to Asset Manager.