Support redirecting to asset preview URLs
Currently the route for assets assumes that it ends with a file
extension (for example `.pdf`). In the case of the preview URLs (which
end with `/preview`) it thinks the format of the page is `.pdf/preview`.
This gets URL encoded into `.pdf%2Fpreview` by Rails and so the redirect
doesn't work.

Rather than handling the file extension special, this PR changes it to
accept any path containing any number of file extensions or slashes in
no particular format.